opt/md_tracker_search_provider/requirements.txt
pygobject
dbus-python
